refactor
makeScreenshots moved to SuntimesScreenshots
